Treesame-Commit-Id: 303d7b4f6a8c9c4144decb4dfbdc97eb3cba0e0f

+++ db/migrate/20210901191217_add_replica_identity_for_pace_plans.rb
@@ -21,11 +21,15 @@ class AddReplicaIdentityForPacePlans < ActiveRecord::Migration[6.0]
   disable_ddl_transaction!
 
   def up
+    return unless defined?(PacePlan)
+
     add_replica_identity "PacePlan", :root_account_id, 0
     remove_index :pace_plans, column: :root_account_id, if_exists: true
   end
 
   def down
+    return unless defined?(PacePlan)
+
     add_index :pace_plans, :root_account_id, algorithm: :concurrently, if_not_exists: true
     remove_replica_identity "PacePlan"
   end
